requirements.txt
pandas>=1.5.0
matplotlib>=3.5.0
numpy>=1.21.0
openai>=1.0.0
python-dotenv>=0.19.0
yfinance>=0.2.0
requests>=2.28.0